recovery: root:Add multidisabler script
Most of the users are still using FBE V1 and complains about decryption. The migration to V2 is a difficult process due to the lack of support in the stock & custom kernel and the carelessness of other developers and users. Add multidisabler script which disables encryption completely for FBE V1.

Without this script, most of the users will continue to use an unofficial and unmaintained version of  TWRP.

Change-Id: I8ecde4e8858de61f564333a060302af7cd4c145b
Signed-off-by: ata-kaner <ata.kaner03@gmail.com>